+++ sktime/transformers/summarise/__init__.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python3 -u
+# coding: utf-8
+
+__author__ = ["Markus Löning"]
+
+from sktime.transformers.summarise._feature_extraction import DerivativeSlopeTransformer
+from sktime.transformers.summarise._feature_extraction import PlateauFinder
+from sktime.transformers.summarise._feature_extraction import RandomIntervalFeatureExtractor
+from sktime.transformers.summarise._tsfresh import TSFreshFeatureExtractor
+from sktime.transformers.summarise._tsfresh import TSFreshRelevantFeatureExtractor
