+++ sktime/utils/data_container.py
@@ -230,9 +230,20 @@ def get_time_index(X):
     return time_index
 
 
-def convert_data(X):
-    columns = []
+def from_nested_to_long(X):
+    """Convert nested dataframe to long dataframe
 
+    Parameters
+    ----------
+    X : pd.DataFrame
+        nested dataframe
+
+    Returns
+    -------
+    Xt : pd.DataFrame
+        long dataframe
+    """
+    columns = []
     for i in range(len(X.columns)):
         df = tabularise(X.iloc[:, i])
         df = df.reset_index()
